+++ src/components/MyReservations/ReservationList.tsx
@@ -2,7 +2,7 @@ import ReservationCard from "./ReservationCard";
 import IconEmpty from "@/assets/svgs/ic_empty.svg";
 import Loading from "@/src/components/Loading";
 import { useGetMyReservations } from "@/src/queries/useMyReservations";
-import { ReservationStatus } from "@/src/types/my-reservations";
+import { ReservationStatus } from "@/src/types/myReservations";
 
 const ReservationList = ({ status }: { status: ReservationStatus }) => {
   const { data, ref, isLoading, isError } = useGetMyReservations({
