+++ src/queries/useMyReservations.ts
@@ -2,13 +2,13 @@ import {
   UpdateMyReservation,
   createMyReservationReview,
   listMyReservations,
-} from "@/src/services/my-reservations";
+} from "@/src/services/myReservations";
+import { GetMyReservations } from "@/src/types/my-reservatios-responses";
 import {
   CreateMyReservationReviewProps,
   ListMyReservationsProps,
   UpdateMyReservationProps,
-} from "@/src/types/my-reservations";
-import { GetMyReservations } from "@/src/types/my-reservatios-responses";
+} from "@/src/types/myReservations";
 import {
   infiniteQueryOptions,
   useInfiniteQuery,
